Use `dask_ml.GridSearchCV` instead of the `sklearn.model_selection.GridSearchCV`
Scikit-learn `GridSearchCV` relies on `joblib.Parallel` to choose the
back-end.

Dask ML does also propose a `GridSearchCV` that performs a similar
search but without using `joblib.Parallel`: it directly creates a Dask
graph that is submitted to the `Scheduler` using the `Client`.

See:
[Documentation](https://ml.dask.org/modules/generated/dask_ml.model_selection.GridSearchCV.html#dask-ml-model-selection-gridsearchcv)
[Code](https://github.com/dask/dask-ml/blob/master/dask_ml/model_selection/_search.py#L1168)

As this latter interface is internally using less fixtures as it does
not need to adapt to `joblib.Parallel` contract, using it may resolve
some of the problems encountered with scikit-learn `GridSearchCV`.