Change the batch size to the number of tasks to execute
The execution is hanging when dispatching tasks.

We can try to just submit one task and see how this goes.
`Parallel.batch_size` (that default to 1 if not specified)
is the number to change accordingly.

This commit change its values to the total number of tasks.

Misc.: it also fixes names of parameters in `fit_and_score_estimator`.
(this was changed in another branch but not on this branch base).